Broke roboBrain out into its own cpp file
